Move scheduler watchdog models into core
