+++ docs/internal/testing/checklists/architecture-improvements.md
@@ -65,6 +65,7 @@ Issue #51 init benchmark baseline evidence (2026-04-28):
 - [x] `ARCH-RTCORE-29` Move bytecode type-table `SIZEOF` evaluation into `trust-runtime-core` without adding a HIR dependency to the pure core path, keep the HIR-gated `SizeOfError` mapper in the runtime compatibility module, and validate focused core/runtime `SIZEOF` gates plus clippy, full-map doctor, and diagram drift.
 - [x] `ARCH-RTCORE-30` Move VM const-pool decoding into `trust-runtime-core`, preserve `trust-runtime::runtime::vm::const_pool` as a compatibility re-export, replace `std` dependencies with `core`/`alloc`, and validate primitive/enum/alias/subrange const decoding plus focused bytecode VM const-error gates.
 - [x] `ARCH-RTCORE-31` Close Phase 5 VM execution movement by leaving `debug_map`, `dispatch_refs`, `local_init`, `dispatch`, `call`, and `register_ir` host-owned for their debug/source lookup, runtime storage, initializer, instruction-loop, call-dispatch, lowering/profile/tier1, and host-callback dependencies; Phase 5 closeout passes core VM tests, `bytecode_vm_core`, register-IR parity, bytecode VM differential, and full-map doctor gates.
+- [x] `ARCH-RTCORE-32` Start Phase 6 model movement by moving `ResourceState` plus watchdog/retain-mode/fault-policy records into `trust-runtime-core`, preserving `trust-runtime::scheduler::ResourceState` and `trust-runtime::watchdog::*` compatibility paths, keeping host-thread scheduler services in `trust-runtime`, and validating core/no-default, scheduler-state, watchdog behavior-lock, clippy, full-map doctor, and diagram drift gates.
 - [x] `ARCH-FULL-01` Add `full-architecture-refactor-program-checklist.md` as the umbrella board that states the runtime-core split is one branch of the full SOLID/KISS and zero-silent-bug program, not the whole program.
 - [x] `ARCH-FULL-02` Add `architecture-doctor-full-map-execution-checklist.md` so `architecture-doctor --full-map` becomes a concrete prerequisite for enforcing workspace, runtime-core, CLI, host-surface, KISS, API, and diagram rules.
 - [x] `ARCH-FULL-03` Add `hir-mutation-hardening-execution-checklist.md` for the F2 silent-bug gap: `symbol_import`, `type_check::const_eval`, and aggregate initializer validation mutation survivors.
